requirements.txt
streamlit
pyperclip
pandas
python-dotenv